+++ TheGuideToTheNewEden/TheGuideToTheNewEden.WinUI/TheGuideToTheNewEden.WinUI/ViewModels/Map/IntelToolViewModel.cs
@@ -45,8 +45,8 @@ namespace TheGuideToTheNewEden.WinUI.ViewModels.Map
         public List<MapIntelMsg> AllIntelMsgs { get; set; }
 
         private MapCanvas _mapCanvas;
-        private Dictionary<int, MapData> _systemDatas;
-        private Dictionary<int, SovData> _sovDatas;
+        private Dictionary<long, MapData> _systemDatas;
+        private Dictionary<long, SovData> _sovDatas;
         private DispatcherTimer _timer;
 
 
@@ -61,7 +61,7 @@ namespace TheGuideToTheNewEden.WinUI.ViewModels.Map
             AllIntelMsgs = new List<MapIntelMsg>();
             RefreshChannel();
         }
-        public void Init(MapCanvas mapCanvas, Dictionary<int, MapData> systemDatas, Dictionary<int, SovData> sovDatas)
+        public void Init(MapCanvas mapCanvas, Dictionary<long, MapData> systemDatas, Dictionary<long, SovData> sovDatas)
         {
             _mapCanvas = mapCanvas;
             _systemDatas = systemDatas;
